move console commands to ConsoleCommands/
